Consolidate account creation and login forms (#3909)
* consolidate forms

* translations
