+++ site/gatsby-site/i18n/locales/es/login.json
@@ -2,6 +2,9 @@
   "Login": "Ingresar",
   "Sign up": "Registrarse",
   "Sign Up": "Registrarse",
+  "Notification Emails": "Correos de notificación",
+  "Continue with email": "Continuar con el correo electrónico",
+  "Enter your email to sign in. We'll create an account for you if you don't have one yet.": "Introduce tu correo electrónico para iniciar sesión. Crearemos una cuenta para ti si aún no tienes una.",
   "Log out": "Salir",
   "Password": "Contraseña",
   "Confirm password": "Confirmar contraseña",
@@ -17,8 +20,6 @@
   "Enter new password": "Ingresa la nueva contraseña",
   "Confirm new password": "Confirma la nueva contraseña",
   "Succesfully updated password, click here to ": "Se ha actualizado la contraseña, haz clic aquí para ",
-  "Don't have an account?": "¿No tienes una cuenta?",
-  "Already have an account?": "¿Ya tienes una cuenta?",
   "Verification email sent to {{email}}": "Email de verificación enviado a {{email}}",
   "Your Account": "Tu cuenta",
   "Subscribe": "Suscribirse",
