+++ site/gatsby-site/i18n/locales/es/translation.json
@@ -301,10 +301,8 @@
   "AIID incidents list": "Lista de incidentes de AIID",
   "Find AI related incidents and reports": "Encuentra incidentes e informes relacionados con la IA",
   "Explore AIID's articles": "Explora los artículos de AIID",
-  "Sign up for an account": "Regístrate para obtener una cuenta",
-  "AIID - Sign Up": "AIID - Registrarse",
-  "AIID - Login": "AIID - Iniciar sesión",
-  "Login to your account": "Inicia sesión en tu cuenta",
+  "AIID - Sign in or sign up": "AIID - Iniciar sesión o registrarse",
+  "Sign in to your AIID account or create a new one": "Inicia sesión en tu cuenta de AIID o crea una nueva",
   "Related News Digest": "Resumen de noticias relacionadas",
   "This summary features AI-related news, identified and matched to our database via NLP for textual similarity, including potential incidents and highlighted harm keywords.": "Este resumen presenta noticias relacionadas con la IA, identificadas y emparejadas con nuestra base de datos mediante PNL para similitud textual, incluyendo incidentes potenciales y palabras clave de daño destacadas.",
   "Website's list of incidents and reports.": "Lista de incidentes e informes del sitio web.",
