+++ site/gatsby-site/i18n/locales/fr/login.json
@@ -2,6 +2,9 @@
   "Login": "Connexion",
   "Sign up": "Enregistrement",
   "Sign Up": "Enregistrement",
+  "Notification Emails": "E-mails de notification",
+  "Continue with email": "Continuer avec l'e-mail",
+  "Enter your email to sign in. We'll create an account for you if you don't have one yet.": "Saisissez votre adresse e-mail pour vous connecter. Nous créerons un compte pour vous si vous n'en avez pas encore.",
   "Log out": "Déconnexion",
   "Password": "Mot de passe",
   "Confirm password": "Confirmer mot de passe",
@@ -17,8 +20,6 @@
   "Enter new password": "Entrez le nouveau mot de passe",
   "Confirm new password": "Confirmez le nouveau mot de passe",
   "Succesfully updated password, click here to ": "Le mot de passe a été mis à jour, cliquez ici pour",
-  "Don't have an account?": "Vous n'avez pas de compte ?",
-  "Already have an account?": "Avez-vous déjà un compte ?",
   "Your Account": "Votre compte",
   "Subscribe": "S'abonner",
   "Subscribe to Major Updates": "S'abonner aux mises à jour majeures",
