+++ site/gatsby-site/i18n/locales/fr/translation.json
@@ -289,10 +289,8 @@
   "AIID incidents list": "Liste des incidents AIID",
   "Find AI related incidents and reports": "Trouver des incidents et des rapports liés à l'IA",
   "Explore AIID's articles": "Explorer les articles de l'AIID",
-  "Sign up for an account": "Créer un compte",
-  "AIID - Sign Up": "AIID - Créer un compte",
-  "AIID - Login": "AIID - Se connecter",
-  "Login to your account": "Connectez-vous à votre compte",
+  "AIID - Sign in or sign up": "AIID - Se connecter ou s'inscrire",
+  "Sign in to your AIID account or create a new one": "Connectez-vous à votre compte AIID ou créez-en un nouveau",
   "Related News Digest": "Résumé des actualités connexes",
   "This summary features AI-related news, identified and matched to our database via NLP for textual similarity, including potential incidents and highlighted harm keywords.": "Ce résumé présente des actualités liées à l'IA, identifiées et appariées à notre base de données via le TAL pour une similarité textuelle, incluant des incidents potentiels et des mots-clés de dommage mis en évidence.",
   "Website's list of incidents and reports.": "Liste des incidents et des rapports du site Web.",
