+++ site/gatsby-site/i18n/locales/ja/login.json
@@ -2,6 +2,9 @@
   "Login": "ログイン",
   "Sign up": "新規登録",
   "Sign Up": "新規登録",
+  "Notification Emails": "通知メール",
+  "Continue with email": "メールで続行",
+  "Enter your email to sign in. We'll create an account for you if you don't have one yet.": "ログインするには、メールアドレスを入力してください。アカウントをお持ちでない場合は、新しく作成します。",
   "Log out": "ログアウト",
   "Password": "パスワード",
   "Confirm password": "パスワード確認",
@@ -17,8 +20,6 @@
   "Enter new password": "新しいパスワードを入力してください",
   "Confirm new password": "新しいパスワードを確認してください",
   "Succesfully updated password, click here to ": "パスワードを更新しました。リンクをクリックしてください ",
-  "Don't have an account?": "アカウントがありませんか？",
-  "Already have an account?": "すでにアカウントをお持ちですか？",
   "Verification email sent to {{email}}": "確認メールを{{email}}に送信しました",
   "Your Account": "あなたのアカウント",
   "Subscribe": "購読",
