+++ site/gatsby-site/i18n/locales/ja/translation.json
@@ -564,10 +564,8 @@
   "AIID incidents list": "AIIDインシデントリスト",
   "Find AI related incidents and reports": "AI関連のインシデントと報告を検索",
   "Explore AIID's articles": "AIIDの記事を探す",
-  "Sign up for an account": "アカウントを作成",
-  "AIID - Sign Up": "AIID - サインアップ",
-  "AIID - Login": "AIID - ログイン",
-  "Login to your account": "アカウントにログイン",
+  "AIID - Sign in or sign up": "AIID - ログインまたは新規登録",
+  "Sign in to your AIID account or create a new one": "AIIDアカウントにログインするか、新しいアカウントを作成してください",
   "This summary features AI-related news, identified and matched to our database via NLP for textual similarity, including potential incidents and highlighted harm keywords.": "この要約には、AI関連のニュースが含まれており、テキストの類似性に基づいてデータベースに特定され、一致させられた潜在的なインシデントと強調された危害キーワードが含まれています。",
   "Website's list of incidents and reports.": "ウェブサイトのインシデントと報告のリスト。",
   "Incidents Over Time": "時間の経過とともにインシデント",
