+++ site/gatsby-site/playwright/e2e/signup.spec.ts
@@ -35,11 +35,11 @@ test.describe('Signup', () => {
 
     const signupResponse = page.waitForResponse('**/api/auth/signin/http-email*');
 
-    await page.locator('[data-cy="signup-btn"]').click();
+    await page.locator('[data-cy="login-btn"]').click();
 
     await signupResponse;
 
-    await expect(page.getByText('A sign up link has been sent to new.user@test.com')).toBeVisible();
+    await expect(page.getByText('A sign in link has been sent to new.user@test.com')).toBeVisible();
   });
 
   test('Should display the error toast message if any other sign up error occurs', async ({ page }) => {
@@ -66,7 +66,7 @@ test.describe('Signup', () => {
 
     const signupResponse = page.waitForResponse('**/api/auth/signin/http-email*');
 
-    await page.locator('[data-cy="signup-btn"]').click();
+    await page.locator('[data-cy="login-btn"]').click();
 
     await signupResponse;
